backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
requests==2.32.0
pydantic[email]==2.7.0
python-dotenv==1.0.0
sqlalchemy==1.4.52
alembic==1.13.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0
pyjwt[crypto]==2.8.0
pyotp==2.8.0
gunicorn==20.1.0
pydantic-settings==2.1.0
slowapi==0.1.8
sentry-sdk[fastapi]==1.40.0
email-validator>=2.1.0
prometheus-fastapi-instrumentator==6.1.0
pytest==8.2.2
httpx>=0.24.0
